Create README.md
This software comprised of the basic frontend functionality, using HTML5 and SCSS to preprocess css, and fontawesome CDN for icons as well as images grabbed from diiferent online source, all of which are legally and openly distributed on the net.

however at this stage of the project responsive mobile and some tab size design comprised of just removing certain element from the page with some little tweaks in overall font size to aid the size adjustment of certain elements. Because of time factor all needed element like nav's would be added subsequently and maybe some few javascript dom manipulation if the need arises,

This page would be updated as the changes are made.